Extract test helpers to a separate package (#3)
As a practice exercise before working on reconciler, I wanted to extract one of the test helper utilities (that Paul pulled from jest) into its own package that reconciler will be able to require. This was a great learning exercise for feeling through the way the current Roblox packaging, bundling, and testing ecosystem works.

Next step: Start on reconciler, which may force me to similarly extract the Timeout utility into a peer package, but I'll do that within the single PR.