Port react/packages/shared and minimal react/packages/react (#8)
shared modules ends up getting woven in and out, directly and indirectly, in reconciler and further up the stack.
bringing most/all of these shared utilities now will better allow us to work in parallel across modules.

the React package itself ends up getting imported from and re-exported from Shared and vice versa. The amount of the React package we need to bring over is relatively small to satisfy what Shared ends up importing from it, so the modules/react portion of this PR will decidedly not be as exhaustive and what we bring over from Shared.

There are a couple slices of real work remaining for the shared folder after this change:

* Introduce the missing tests that depend on react-dom and react
* Convert and fix up the rest of ReactComponentStackFrame.roblox.lua to mimic the stack inspection functionality in the original file