Global vars (#18)
Replaces __DEV__ with _G.__DEV__ in most places to keep Luau analysis happy until Luau supports the ability to define valid globals in a definitions file or the like.

This caught a few missing bits and bobs elsewhere, which is encouraging. We plan to come back and find-replace these back to regular globals later when we have proper Luau support.