Noop renderer (#23)
Translate the no-op rendererer, excluding only a few JSX-related functions.

I finished converting it without types, got it to analyze, then realized that the few locally defined object types were pretty manageable to reintroduce, so I did so. This wouldn't check in strict mode, but it works fine in nonstrict and is a much closer translation!

This also includes some small changes to the scheduler. In the tests for the scheduler package, we use a version of the default scheduler that also reexports some members of the mock HostConfig that it's also using. I've adapted this to be a little bit closer (though still not quite the same) to the version being used upstream.