Enabling the rest of WorkLoop (#43)
With the additional locals headroom, enabled more big chunks of the workloop.
Still TODO:
-See if there's an easy extraction to break a newly discovered cycle: RectFiberCommitWork imports captureCommitPhaseError from ReactFiberWorkLoop
-add git commit hash link to top of all newly-ported files

Several tests that only passed in non-DEV mode now pass. Only one still doesn't. With the added console.warn for when errors happen deep in the framework, it should now be much easier to see what the root cause of things are. There's a few unexpected warnings that happen, even though the tests are technically passing.

After this, it looks like we might be ready to implement useState() driven the tests in this same file?