React fiber class component (#44)
Implements the majority of the logic for class components and their lifecycle management.

The tests included here are adapted from tests in react-dom in the upstream repo, so they're not exactly one-to-one. These tests were first adapted in this upstream branch and then ported once a significant amount of them were passing.

There are some other bits of cleanup, including:

* Require console from shared instead of the polyfill dependency directly; this lets us capture errors and warnings in dev mode.
* Fixes some test failures around unmounting work that had been previously mounted; the fix involved using a sentinel value to indicate the need for deletion
* Rotriever dependency on specific hash of Luau Polyfill to avoid devs getting out of sync (to change this, we'll need to change that hash and run rotrieve install)
* Use of Symbol for some of the constants, which improves readability during debugging