Roblox renderer v2 (#54)
Starting to stand up the logic for the Roblox renderer, something much closer to what we'll expect to see in the final project. For now, I've dragged a bunch of placeholder stuff over to try to get a sense of what the interface will look like.

Ultimately, this involves a lot of roughly-translated pieces of ReactDOM, so that we can get certain abstractions that form the public API of ReactDOM, something we'd probably like to mimic either way ("roots" is a good example).

Nearly all of the lifted Roact code brought its unit tests with it (bindings, events, etc.). We may want to promote bindings back up to the reconciler at some point, and maybe even propose them as an upstream feature someday!