Fix roblox-cli analyze not reporting correctly (#69)
Co-authored this with Max, who saw that there were files that didn't seem to be respected by the linting. This turned out to be a side effect of the rearranged project structure introduced in #54, in which the path to the Packages folder was adjusted to match rotriever convention. This project format ended up placing the .robloxrc in the Packages directory above all of our source code, causing large amounts of it to be ignored.

We fixed this by adding .robloxrc files alongside the source of each workspace member; this is a clumsy workaround, as it mixes config in with source code and it risks type errors when RoactAlignment is used as a dependency. However, it's the only configuration that works at the moment. We expect that this use case will be addressed in the future with luau config features or roblox universal project format features, and for now this makes using RoactAlignment as a dependency possible at all.

Finally, once we fixed this, a number of type errors surfaced that were being swallowed in #54. I've fixed all of these errors and left comments for uncertain type translations and work left to do. I was also able to remove a workaround for a strange set of type errors that seemed like they shouldn't have been happening, likely another side effect of the config misbehaviors.