Clean up some test brittleness. (#78)
Throw all errors in ReactErrorBoundaries tests without stacktraces so that it's easier validate the output. Uses the second argument to the error function (https://developer.roblox.com/en-us/api-reference/lua-docs/Lua-Globals):

"void error ( string message, int level = 1 )
Terminates the last protected function called and outputs message as an error message. If the function containing the error is not called in a protected function (pcall), then the script which called the function will terminate. The error function itself never returns and acts like a script error.
The level argument specifies how to get the error position. With level 1 (the default), the error position is where the error function was called. Level 2 points the error to where the function that called error was called; and so on. Passing a level 0 avoids the addition of error position information to the message."

(from https://www.tutorialspoint.com/lua/lua_error_handling.htm)

This change allows some of the work in #77 to rearrange package paths without breaking tests!