Renderer entry point (#118)
Adjusts the way that the entry-point to the reconciler package works. Rather than immediately exposing its contents, the reconciler package exposes a function that can be called with a provided config object to dynamically inject the proper host config. This mimic the way that test/noop/shallow renderers were already mocking the reconciler entrypoint (and actually eliminates the need for a couple of mocks).

This should mean that the ReactRoblox renderer can now be depended upon as you would react-dom and it should also make the RoactCompat package fully usable as a drop-in replacement for old Roact.

A few additional notes:

* There are still unresolved deviations that are present even when using the RoactCompat package, including state initialization, the shape of Context.Consumer's api, etc.
* This initialization step must happen before the rest of ReactReconciler's contents are required. This works fine for most external use cases which only need ReactRoblox anyways, and is supported by internal tests that will reset modules in between (and thus un-initialize the host config)
* Downstream users who want to use multiple renderers (e.g. some tests use TestRenderer, others use ReactRoblox) will need to also do the module resetting dance. If we can't improve this workaround, we may need to document this in detail somewhere