LUAFDN-328: Include component name in lifecycle deprecation warning (#120)
Right now, the lifecycle deprecation warning doesn't give information about where it was being used, which is unhelpful when the user is trying to go replace it with the correct method.

This adds the component name into the warning.
We may also want to find a way to make the warning clickable to open to the component.