React roblox planning (#141)
A step towards closing out LUAFDN-382. I've made a few small touchup change that seemed uncontroversial and written out some documentation on the more complicated TODOs for ReactRoblox.

They boil down to:

Keys provided to elements in Roact are applied to the Name field of the resulting Instance; ReactRoblox does not currently do that.
Refs in Roact are implemented in terms of bindings, and certain ref use cases depend on those internals being the same
Text instances are unimplemented. I'm fine with keeping it that way, but it's a deviation to call out.
findDOMNode is unimplemented. It seems more dangerous than useful.
Legacy render and unmountComponentAtNode functions aren't implemented in roact-alignment, but are expected to be deprecated soon in upstream
Some reconciler batching and scheduling logic is exposed with unstable_ prefixes
Hydration is unimplemented
unstable_createEventHandle appears to be an upcoming way of managing event listeners (replacing a useEvent hook); it's not clear if it applies to ReactRoblox in anyway
A few other top-level apis with unstable prefixes, most of which don't seem necessary to port
React has special behavior for input, select, option, and textarea tags that helps them respect the intended source of truth; it's possible that we should implement something similar for at least TextBox. Not sure if there's anything else it would apply to
Roact.Change and Roact.Event have been pulled in from Roact, but we might want to think about if there's a way to make their logic more generic. For what it's worth, upstream does not seem concerned about whether they're generic or not.