Update mount return type to RoactHandle (to match unmount param type) (#265)
Closes #264.

The current inferred return type for `RoactTree.mount` does not match
the param type accepted by `RoactTree.unmount`, leading to analyzer
errors when strict mode is enabled.

The issue boils down to the key param to mount being of type `string?`.
It is later assigned a default string value if `nil` but Luau is not able to
update the type of the existing local var from `string?` to `string`.

We can work around this with an explicit cast. The return type of mount
is also made explicit (reusing the existing `RoactHandle` type).