LUAFDN-803 - Tag Support (#276)
Closes LUAFDN-803.

Introduces an implementation of tags on host components that provides a declarative, React-y interface to the Roblox CollectionService API.

This approach accepts a list of tags as a single, space-delimited string. This approach should balance the ergonomics of a list-driven approach with practical performance concerns around the expense of list diffing. Since tags will most often be applied statically (the same set of tags for every render of a given component), we can leverage string identity equality to avoid diffing the same set of tags.