LUAFDN-1148 Port uncacheFiberNode logic from ReactNative (#308)
ReactRobloxComponentTree uses a similar mechanism to ReactNativeComponentTree for tracking which fiber node corresponds to each host instance, by maintaining a set of internal maps which are set when a new fiber node is created.

However our implementation can leak memory because we do not provide a corresponding uncacheFiberNode function, which is present in ReactNativeComponentTree. This leads to the tables used for tracking these instances to continuously grow in size when new elements are created, never being cleared.

This can be fixed by a 1:1 port of the uncacheFiberNode logic from ReactNative, when nodes are no longer used the corresponding entry will be removed from the table.