LUAFDN-1130 - Friendlier hook deps (#305)
Adds a different mechanism for iterating through dependency arrays. I like this approach because while it has some extra iteration, it doesn't introduce any new table creation so it shouldn't hurt our GC footprint.

This also disables the warning that would be shown when trailing nils cause arrays to be different lengths between renders; of course it also means that we won't detect the valid scenarios in which this could happen.