Add 'return nil' to several functions to resolve Luau type errors (#328)
We are making some improvements to Luau's type checking of the number of values returned by a function. Today, there are some cases where Luau fails to report a diagnostic in certain cases where a function returns fewer values than the caller expects. We are making improvements to ensure we consistently report errors.

This work has exposed an existing deficiency with Luau's type analysis around xpcall usage. xpcall can return two completely different things, depending on whether the operation was successful or not. The Luau type checker is not currently able to model this, and so it currently uses only the "success" path to determine the type.

This change makes a couple of tweaks to SchedulerHostConfig.default.lua to appease Luau given this limitation of its type checking of xpcall:

We modify doWork to explicitly return nil (so that it returns one value instead of zero, to match the number of values returned by describeError).
In the error path, we cast the returned value to any.